+++ src/components/select/components/focus-cover.js
@@ -1,0 +1,8 @@
+import ReactDOM from 'react-dom';
+
+export default function FocusCover({children}) {  
+  return ReactDOM.createPortal(
+    children,
+    document.body
+  )
+}