+++ src/components/select/select.js
@@ -1,6 +1,8 @@
 import React from 'react'
 import ManagedInput from '../input/input'
+import FocusCover from './components/focus-cover'
 import * as handlers from './handlers/handlers'
+import { componentDidUpdate } from './lifecycle/lifecycle'
 
 export default class ManagedSelect extends ManagedInput {
   constructor(props) {
@@ -10,9 +12,11 @@ export default class ManagedSelect extends ManagedInput {
     this.optionsClass = this.classNames.options || 'options'
     this.optionClass = this.classNames.option || 'option'
     this.markClass = this.classNames.mark || 'mark'
+    this.componentDidUpdate = componentDidUpdate
     this.onClickHandler = handlers.onClickHandler.bind(this)
     this.onKeyDownHandler = handlers.onKeyDownHandler.bind(this)
     this.onBlurHandler = handlers.onBlurHandler.bind(this)
+    this.onFocusCoverClickHandler = handlers.onFocusCoverClickHandler.bind(this)
     this.onSelectHandler = handlers.onSelectHandler.bind(this)
     this.onOptionMouseOver = handlers.onOptionMouseOver.bind(this)
     this.onFocusCoverZIndex = this.props.onFocusCoverZIndex || 2
@@ -30,65 +34,77 @@ export default class ManagedSelect extends ManagedInput {
 
     return (
       <div
-      style={{ display }}
-      className={this.containerClass + decorator}>
+        style={{ display }}
+        className={this.containerClass + decorator}
+      >
         <div
-        onMouseOver={this.onMouseOverHandler}
-        onMouseLeave={this.onMouseLeaveHandler}>
+          onMouseOver={this.onMouseOverHandler}
+          onMouseLeave={this.onMouseLeaveHandler}
+        >
           <div onClick={this.onClickHandler}>
             <div className={this.labelClass + decorator}>
               {label}
             </div>
             <div className={this.inputClass + decorator}>
-              {
-              (options.find(option => (option.value === value)) || {}).text
-
-              }
+              {(options.find(option => (option.value === value)) || {}).text}
               <div className={this.markClass + decorator} />
             </div>
           </div>
           {
-          this.state.focus && this.isMobile && !this.disableOnFocusCover &&
-          <div
-          onClick={() => this.onBlurHandler(undefined, undefined, true)}
-          style={{ position: 'fixed', top: 0, right: 0, bottom: 0, left: 0, zIndex: this.onFocusCoverZIndex }}>
-          </div>
+          (this.state.focus && !this.disableOnFocusCover) &&
+          <FocusCover>
+            <div
+              onClick={this.onFocusCoverClickHandler}
+              style={{
+                position: 'fixed',
+                top: 0,
+                right: 0,
+                bottom: 0,
+                left: 0,
+                zIndex: this.onFocusCoverZIndex
+              }}
+            />
+          </FocusCover>
           }
           {
           this.state.focus && !this.disableSelectOptions &&
           <div
-          style={{ zIndex: 999 }}
-          ref={this.options}
-          className={this.optionsClass + decorator}>
+            style={{ zIndex: 999 }}
+            ref={this.options}
+            className={this.optionsClass + decorator}
+          >
             {
             options.map((option, index) => (
             <div
-            key={`key-proxy-${option.value}-${index}`}
-            onMouseOver={() => this.onOptionMouseOver(index)}
-            onClick={() => this.onSelectHandler(option)}
-            className={this.optionClass + decorator + `${this.state.preSelected === index ? ' preselected' : ''}`}>
+              key={`key-proxy-${option.value}-${index}`}
+              onMouseOver={() => this.onOptionMouseOver(index)}
+              onClick={() => this.onSelectHandler(option)}
+              className={this.optionClass + decorator + `${this.state.preSelected === index ? ' preselected' : ''}`}
+            >
               {option.text}
             </div>))
             }
           </div>
           }
           <select
-          ref={this.input}
-          id={this.id}
-          value={value}
-          readOnly={true}
-          style={{ position: 'absolute', left: -10000 }}
-          disabled={this.state.disabled}
-          onFocus={this.onFocusHandler}
-          onKeyDown={this.onKeyDownHandler}
-          onBlur={this.onBlurHandler}>
+            ref={this.input}
+            id={this.id}
+            value={value}
+            readOnly={true}
+            style={{ position: 'absolute', left: -10000 }}
+            disabled={this.state.disabled}
+            onFocus={this.onFocusHandler}
+            onKeyDown={this.onKeyDownHandler}
+            onBlur={this.onBlurHandler}
+          >
             {
             options.map((option, index) => (
-            <option
-            key={`key-original-${option.value}-${index}`}
-            value={option.value}>
-              {option.value}
-            </option>))
+              <option
+                key={`key-original-${option.value}-${index}`}
+                value={option.value}
+              >
+                {option.value}
+              </option>))
             }
           </select>
         </div>